+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/taxonkit/v0.16.0/total.svg)](https://github.com/shenwei356/taxonkit/releases/tag/v0.16.0)
     - `taxonkit create-taxdump`:
         - TaxIds are generated from the hash value of "rank+taxon_name", this allowed duplicated names with different ranks (common in GTDB). [#92](https://github.com/shenwei356/taxonkit/issues/92).
+    - `taxonkit taxid-changelog/create-taxdump`:
+        - Add notes about the possible error records in taxid-changelog built from taxonomic data created by `create-taxdump`. [#91](https://github.com/shenwei356/taxonkit/issues/91)
+        - Dont't worry, a single version of taxonomic data created by `taxonkit create-taxdump` has no problem, it's just the changelog might not be perfect.
     - `taxonkit lca`:
         - new flag `-K/--keep-invalid`: print the query even if no single valid taxid left. [#89](https://github.com/shenwei356/taxonkit/issues/89)
 - [TaxonKit v0.15.1](https://github.com/shenwei356/taxonkit/releases/tag/v0.15.1)